refactor(models): move model structs and errors into dedicated module
- Extracted `LanguageModel`, `RerankingModel`, and `EmbeddingModel` structs and their error types from `lib.rs` into new files under `models/`.
- Updated imports and usage throughout the codebase to reference the new `models` module.
- Moved related trait definitions (`GeneratesText`, `GeneratesObject`, `Reranks`, `ReranksStructuredData`, `Embeds`) to their respective response generator modules.
- Centralized `AIProvider` enum and its implementations in `providers/mod.rs` for better separation of concerns.
- No functional changes; this is an internal code organization and modularization improvement.
